[RecoPixelVertexing/PixelVertexFinding] merged header of PixelVertexProducer into .cc
